% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_info.R
\name{image_info}
\alias{image_info}
\title{Fetch information about an image, given its image ID}
\usage{
image_info(id, verbose = sbdi_config()$verbose)
}
\arguments{
\item{id}{character: IDs of images (e.g. as returned by \code{\link{occurrences}} 
in the imageUrl column). Each ID will be of a format something like "84654e14-dc35-4486-9e7c-40eb2f8d3faa"}

\item{verbose}{logical: show additional progress information? [default is set by sbdi_config()]}
}
\value{
A data.frame with one row per \code{id}, and at least the columns 
imageIdentifier and imageURL
}
\description{
Note that there is currently no web service that provides image information, 
and so we are scraping results from pages of the form https://images.bioatlas.se/image/details?imageId=id. 
This web scraping may be fragile, and will be replaced by a web-service-based function when one becomes available.
}
\examples{
\dontrun{
## Using IDs returned from occurrences() function

image(id = c("f31e5f0e-f964-4bc3-b8f3-78f2ad520563",
  "39091d30-3016-417f-8860-b7c2a8d37942"))
}
}
\seealso{
\code{\link{sbdi_config}} \code{\link{occurrences}}
}
