% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbn_cache_filename.R
\name{nbn_cache_filename}
\alias{nbn_cache_filename}
\title{Returns the name of the cache file associated with the given URL. Note that this file 
may not actually exist, this function just provides the mapping from URL to filename}
\usage{
nbn_cache_filename(url)
}
\arguments{
\item{url}{string: the URL}
}
\value{
string: the file path and name
}
\description{
Returns the name of the cache file associated with the given URL. Note that this file 
may not actually exist, this function just provides the mapping from URL to filename
}
\examples{
nbn_cache_filename("https://records-ws.nbnatlas.org/index/fields")

}
\references{
\url{https://api.nbnatlas.org/}
}
\seealso{
\code{nbn_config} for cache settings, particularly the cache directory
}
