% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbn_lists.R
\name{nbn_lists}
\alias{nbn_lists}
\title{Find NBN species lists}
\usage{
nbn_lists(guid, offset = 0, max = 500, verbose = nbn_config()$verbose)
}
\arguments{
\item{guid}{string: (optional) if provided, return only lists in which this GUID appears}

\item{offset}{integer: the number of lists to skip. This supports paging}

\item{max}{integer: the maximum number of lists to return. This supports paging}

\item{verbose}{logical: show additional progress information?}
}
\value{
data.frame of list name and other details
}
\description{
Find NBN species lists
}
\examples{
\dontrun{
 ## lists that include the giant African snail Achatina fulica
 ##  (which is a notifiable pest species in some states)
 l <- nbn_lists(search_guids("Achatina fulica")$guid)
}

}
\references{
\url{https://lists.nbnatlas.org} and the associated web services at \url{https://lists.nbnatlas.org/ws}
}
\seealso{
\code{\link{nbn_list}}
}
