% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_filter.R
\name{pick_filter}
\alias{pick_filter}
\title{List and pick existing filters}
\usage{
pick_filter(type = NULL)
}
\arguments{
\item{type}{string: (optional) type of filter to create query string for. Options are \code{c("resource", "specieslist" , "layer")}. Species lists are waiting development from the API side.}
}
\value{
a string ready to be places in the argument \sQuote{fq} in function \code{\link{occurrences}}.
}
\description{
Lists and lets the user interactivelly pick a filter from lists.
}
\examples{
\dontrun{
fq_str <- pick_filter("resource") #"data_resource_uid:dr5"
x <- occurrences(taxon="genus:Accipiter", 
                  download_reason_id=10,
                  fq=fq_str)
}
}
\seealso{
\code{\link{occurrences}} for download reasons; \code{\link{sbdi_config}}
}
