% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Scatterplot of environmental parameters associated with species occurrence data. Interactive plotly plot of environmental variables and species occurence data}
\usage{
scatterplot(taxon, ...)
}
\arguments{
\item{taxon}{string: (optional) query of the form field:value (e.g. "genus:Macropus") or a free text search (e.g. "macropodidae"). Note that
a free-text search is equivalent to specifying the "text" field (i.e. \code{taxon="Alaba"} is equivalent to \code{taxon="text:Alaba"}. 
The text field is populated with the taxon name along with a handful of other commonly-used fields, and so just specifying your target
taxon (e.g. taxon="Alaba vibex") will probably work. 
However, for reliable results it is recommended to use a specific field where possible (see \code{nbn_fields("occurrence_indexed")}
for valid fields). It is also good practice to quote the taxon name if it contains multiple words, for example
\code{taxon="taxon_name:\"Alaba vibex\""}}

\item{\dots}{: other options passed to occurences()}
}
\value{
Data frame of occurrence results, with one row per occurrence record. The columns of the dataframe will depend on the requested fields. The data frame is plotted with ggplot and output stored as pdf, an interactive ggplot using plotly is displayed on-screen.
}
\description{
Makes a search using \code{\link{occurrences}}and plots environmental variables associated with species presence against each other. Alternatively, makes a search using \code{\link{occurrences}}and plots species occurence data e.g frequencies against environmental variables. Note that there is a limit of 500000 records per request when using \code{method="indexed"}. Use the \code{method="offline"} for larger requests. For small requests, \code{method="indexed"} likely to be faster.
}
\examples{
\dontrun{ 
scatterplot(taxon="Ectocarpus siliculosus")
}
}
\references{
\itemize{
\item Associated NBN web service for record counts: \url{https://api.nbnatlas.org/#ws3}
\item Associated NBN web service for occurence downloads: \url{https://api.nbnatlas.org/#ws4}
\item Field definitions: \url{https://docs.google.com/spreadsheet/ccc?key=0AjNtzhUIIHeNdHhtcFVSM09qZ3c3N3ItUnBBc09TbHc}
\item WKT reference: \url{http://www.geoapi.org/3.0/javadoc/org/opengis/referencing/doc-files/WKT.html}
}
}
\seealso{
\code{\link{sbdi_reasons}} for download reasons; \code{\link{sbdi_config}}
}
