% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBDI.R
\docType{package}
\name{SDBI4R}
\alias{SDBI4R}
\title{\pkg{SBDI4R}}
\description{
REWRITE This project enables the R community to access data and tools hosted by the 
SBDI Atlas. The goal of the project is to enable basic species and related
information to be queried and output in standard formats for R. 
SBDI4R is based around the extensive web services provided 
by the Atlas; see the API link below.
}
\references{
\url{https://api.bioatlas.se/}
}
