% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swe.R
\docType{data}
\name{swe}
\alias{swe}
\title{Basic Sweden polygon for visualization
A basic polygon dataset in SpatialPolygons (package sp) representing
Sweden.}
\format{
SpatialPolygons (package sp) CRS: EPSG:3006
}
\source{
Statistikmyndigheten SCB \url{https://www.scb.se/hitta-statistik/regional-statistik-och-kartor/regionala-indelningar/digitala-granser}
}
\usage{
data(swe)
}
\description{
Basic Sweden polygon for visualization
A basic polygon dataset in SpatialPolygons (package sp) representing
Sweden.
}
\keyword{datasets}
