% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swe.R
\docType{data}
\name{sweWGS84}
\alias{sweWGS84}
\title{Basic Sweden polygon for visualization
A basic polygon dataset in SpatialPolygons (package sp) representing
Sweden.}
\format{SpatialPolygons (package sp) CRS: EPSG:4326}
\source{
Statistikmyndigheten SCB \url{https://www.scb.se/hitta-statistik/regional-statistik-och-kartor/regionala-indelningar/digitala-granser}
}
\usage{
data(sweWGS84)
}
\description{
Basic Sweden polygon for visualization
A basic polygon dataset in SpatialPolygons (package sp) representing
Sweden.
}
\keyword{datasets}
