% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBDI4R.R
\docType{package}
\name{SDBI4R}
\alias{SDBI4R}
\title{\pkg{SBDI4R}}
\description{
This project enables the R community to access data and tools hosted by the 
Swedish Biodiversity Data Infrastructure (SBDI). The goal of the project is 
to enable basic species and related information to be queried and output in 
standard formats for R. 
SBDI4R is based around the extensive web services provided 
by the BioAtlas Sweden; see the API link below.
}
\references{
\url{https://api.bioatlas.se/}
}
